공지사항,FAQ 리포지토리 작업

+++ cashnamu_v2/src/main/java/cashnamu/cashnamu_v2/www/modules/information/faq/service/Faq.java
@@ -32,7 +32,7 @@ import lombok.Setter;
 public class Faq {
 
 	@Id @GeneratedValue
-	private int idx;
+	private Long idx;
 	
 	@Column(length = 100)
 	private String title;
